chore: remove produce block apis no longer part of spec
